Removing all code except for isAccordion from BaseAccordion
